2019-11-07 02:53 提高评测可靠性（事务管理） 将所有Mybatis的Map返回值改为类

+++ src/main/java/com/example/myproject2/entity/RunCode.java
@@ -15,14 +15,14 @@ public class RunCode {
     private String runCodeFile;
     private String codeType;
     private String testDataPath;
-    private int testDataSerial;
+    private short testDataSerial;
     private String status;
     private short runTime;
     private int runMemory;
 
     public RunCode() {
     }
-    public RunCode(int submitCodeId, String runCodeFile, String codeType, String testDataPath, int testDataSerial) {
+    public RunCode(int submitCodeId, String runCodeFile, String codeType, String testDataPath, short testDataSerial) {
         this.runCodeId = runCodeId;
         this.submitCodeId = submitCodeId;
         this.runCodeFile = runCodeFile;
@@ -74,11 +74,11 @@ public class RunCode {
         this.testDataPath = testDataPath;
     }
 
-    public int getTestDataSerial() {
+    public short getTestDataSerial() {
         return testDataSerial;
     }
 
-    public void setTestDataSerial(int testDataSerial) {
+    public void setTestDataSerial(short testDataSerial) {
         this.testDataSerial = testDataSerial;
     }
